+++ test/authorizer/authorizer.feature.test.ts
@@ -480,7 +480,7 @@ describe('Feature: methods throw if the authorizer has not yet been authenticate
     beforeAll(() => {
       authenticated.authenticate();
     });
-    describe.each(['getUser', 'getRoles', 'getClient', 'inScope'])(
+    describe.each(['getUser', 'inScope'])(
       'When: invoking %s on the unauthenticated authorizer',
       methodName => {
         test('Then: an error is thrown', () => {
@@ -488,7 +488,7 @@ describe('Feature: methods throw if the authorizer has not yet been authenticate
         });
       }
     );
-    describe.each(['getUser', 'getRoles', 'getClient', 'inScope'])(
+    describe.each(['getUser', 'inScope'])(
       'When: invoking %s on the authenticated authorizer',
       methodName => {
         test('Then: no error is thrown', () => {